Hangwei_BackEnd/hangwei/opinion/requirements.txt
baidu_aip==2.2.13.0
redis==3.2.1
jieba==0.39
gensim==3.7.3
pandas==0.24.2
requests==2.9.1
analysis==0.1.7
seaborn==0.9.0
scikit_learn==0.21.1
